feat: role

+++ controllers/UserController.php
@@ -1,5 +1,4 @@
 <?php
-
 namespace app\controllers;
 
 use Yii;
@@ -7,6 +6,7 @@ use yii\web\Controller;
 use app\models\User;
 use app\models\UserLevel;
 use yii\data\ActiveDataProvider;
+use yii\web\NotFoundHttpException;
 
 class UserController extends Controller
 {
@@ -14,16 +14,30 @@ class UserController extends Controller
 
     public function actionIndex()
     {
+        $query = User::find();
+
+        // Jika pengguna adalah Kepala Bagian atau Karyawan, hanya tampilkan data mereka sendiri
+        if (Yii::$app->user->identity->id_user_level == 2 || Yii::$app->user->identity->id_user_level == 3) {
+            $query->andWhere(['id_user' => Yii::$app->user->identity->id]);
+        }
+
         $dataProvider = new ActiveDataProvider([
-            'query' => User::find(),
+            'query' => $query,
         ]);
 
         return $this->render('index', [
             'dataProvider' => $dataProvider,
         ]);
     }
+
     public function actionCreate()
     {
+        // Hanya Admin yang bisa mengakses aksi create
+        if (Yii::$app->user->identity->id_user_level != 1) {
+            Yii::$app->session->addFlash('error', 'Anda tidak diizinkan untuk menambah data user.');
+            return $this->redirect(['index']);
+        }
+
         $model = new User();
         $userLevels = UserLevel::find()->all();
 
@@ -45,6 +59,11 @@ class UserController extends Controller
 
     public function actionUpdate($id)
     {
+        // Pastikan pengguna hanya dapat mengedit data mereka sendiri jika mereka bukan Admin
+        if (Yii::$app->user->identity->id_user_level != 1 && $id != Yii::$app->user->identity->id) {
+            throw new NotFoundHttpException('Anda tidak diizinkan untuk mengedit data ini.');
+        }
+
         $model = User::findOne($id);
         $userLevels = UserLevel::find()->all();
 
@@ -69,6 +88,11 @@ class UserController extends Controller
 
     public function actionDelete($id)
     {
+        // Pastikan hanya Admin yang dapat menghapus data
+        if (Yii::$app->user->identity->id_user_level != 1) {
+            throw new NotFoundHttpException('Anda tidak diizinkan untuk menghapus data ini.');
+        }
+
         User::findOne($id)->delete();
         Yii::$app->session->addFlash('success', 'Data berhasil dihapus!');
         return $this->redirect(['index']);
@@ -76,6 +100,11 @@ class UserController extends Controller
 
     public function actionView($id)
     {
+        // Pastikan pengguna hanya dapat melihat data mereka sendiri jika mereka bukan Admin
+        if (Yii::$app->user->identity->id_user_level != 1 && $id != Yii::$app->user->identity->id) {
+            throw new NotFoundHttpException('Anda tidak diizinkan untuk melihat data ini.');
+        }
+
         $model = User::findOne($id);
 
         return $this->render('view', [
